website/requirements1.txt
Django==2.2.7
get==2019.4.13
post==2019.4.13
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.0
